Adds Swanlab notification methods
Adds support for Swanlab notifications via Slack.

This change introduces the `swanlab_notification_methods` parameter,
allowing users to specify which notification channels (Lark and/or Slack)
should be used to receive experiment updates. It also includes corresponding
parameters for the Lark and Slack webhook URLs and necessary logic to
register the appropriate callbacks based on the selected notification methods.