[megatron] fix: make bridge exported cloned weights store on CPU
There are so many `clone` in the GPTBridge code, if we make a
copy of the weight on GPU, OOM can easily happen. This PR tries
to address this issue by using cpu clone as a solution, so that
those weight will be cloned to CPU.

This PR also add `torch.cuda.empty_cache` to mitigate the possibility
of OOM during LoRA merge.

Signed-off-by: Hollow Man <hollowman@opensuse.org>